@@ -26,28 +26,39 @@
         }
     },
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.4",
         "ext-json": "*",
-        "keboola/storage-api-client": "^11.2|^12.1"
+        "keboola/storage-api-client": "^12.10",
+        "symfony/http-foundation": "^4.4|^5.2",
+        "symfony/validator": "^4.4|^5.2"
     },
     "require-dev": {
-        "php-parallel-lint/php-parallel-lint": "^1.0",
-        "phpunit/phpunit": "^5.6"
+        "keboola/coding-standard": "^13.0",
+        "php-parallel-lint/php-parallel-lint": "^1.3",
+        "phpstan/phpstan": "^0.12.99",
+        "phpstan/phpstan-phpunit": "^0.12.20",
+        "phpunit/phpunit": "^9.5",
+        "symfony/dotenv": "^5.4"
     },
     "scripts": {
-        "tests": "phpunit --coverage-clover build/logs/clover.xml --coverage-xml=build/logs/coverage-xml --log-junit=build/logs/phpunit.junit.xml",         
-        "phplint": "parallel-lint -j 10 --exclude vendor .",
-        "build": [
-            "@phplint",
-            "@tests"
-        ],
+        "phplint": "parallel-lint -j 10 --exclude vendor --exclude bin .",
+        "phpcs": "phpcs --extensions=php src tests",
+        "phpcbf": "phpcbf --extensions=php src tests",
+        "phpstan": "phpstan analyse --no-progress -c phpstan.neon.dist",
+        "phpunit": "phpunit",
         "ci": [
             "@composer validate --no-check-all --strict",
-            "@build"
+            "@phplint",
+            "@phpcs",
+            "@phpstan",
+            "@phpunit"
         ]
     },
     "config": {
         "sort-packages": true,
-        "process-timeout": 3600
+        "process-timeout": 3600,
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
+        }
     }
 }
